+++ src/main/java/org/fenixedu/academic/domain/accounting/postingRules/serviceRequests/PhotocopyRequestPR.java
@@ -49,7 +49,7 @@ public class PhotocopyRequestPR extends PhotocopyRequestPR_Base {
     @Override
     protected Money getAmountForPages(final Event event) {
         final PhotocopyRequestEvent requestEvent = (PhotocopyRequestEvent) event;
-        final int extraPages = requestEvent.getNumberOfPages().intValue() - 1;
+        final int extraPages = requestEvent.getNumberOfPages() - 1;
         return getAmountPerPage().multiply(BigDecimal.valueOf(extraPages < 0 ? 0 : extraPages));
     }
